requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
google-cloud-bigquery==3.14.1
google-cloud-secret-manager==2.17.0
google-cloud-aiplatform==1.39.0
elasticsearch==8.11.1
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0